pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-agent"
version = "0.1.0"
description = "Local-first LLM coding agent with Bedrock and Ollama backends"
readme = "README.md"
requires-python = ">=3.12"

authors = [
  { name = "Aaron McAughan" }
]

dependencies = [
  # AWS
  "boto3>=1.34",
  "botocore>=1.34",

  # Strands
  "strands-agents[ollama]>=0.1.0",
  "strands-agents-tools>=0.1.0",

  # HTTP + config
  "requests>=2.31",
  "python-dotenv>=1.0",

  # CLI niceties (optional but useful)
  "rich>=13.7",

  # Typing / Models
  "pydantic>=2.6",

  # For reading config
  "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
]

[project.scripts]
agent = "agent.cli:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
  "smoke_ollama: end-to-end smoke test for local Ollama backend",
  "smoke_bedrock: end-to-end smoke test for Bedrock backend",
]
